+++ tests/api/removeFolderCept.php
@@ -3,7 +3,7 @@
 
 use Codeception\Util\HttpCode;
 
-require __DIR__ . '/../../src/Helper.php';
+require_once __DIR__ . '/../../src/Helper.php';
 
 $I = new ApiTester($scenario);
 $files_root = realpath(__DIR__ . '/../files') . '/';